Split DSType into its own source file
